requirements.txt
# Coach d'Entraînement Intelligent - Dependencies


# Web Framework
streamlit>=1.28.0

# Data Validation
pydantic>=2.0.0

# API Integrations
garminconnect>=0.2.0
google-api-python-client>=2.100.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Visualization
plotly>=5.17.0
matplotlib>=3.7.0

# AI/ML
mistralai>=0.0.7

# Utilities
python-dotenv>=1.0.0
pytz>=2023.3

# File parsing (for .FIT files if needed)
fitparse>=1.2.0

# HTTP requests
requests>=2.31.0

# Date/Time
python-dateutil>=2.8.2

# Optional: Weather API
pyowm>=3.3.0

# Testing (development)
pytest>=7.4.0
pytest-cov>=4.1.0
